+++ babel-test-hook.js
@@ -1,11 +1,22 @@
-require('babel-register')({
-  presets: ['env', 'stage-2', 'react'],
-  ignore: (fileName) => {
+require('@babel/register')({
+  plugins: [
+    ['@babel/plugin-proposal-decorators', { 'legacy': true }],
+    ['@babel/plugin-proposal-class-properties', { 'loose': true }],
+    '@babel/plugin-proposal-export-namespace-from',
+    '@babel/plugin-proposal-function-sent',
+    '@babel/plugin-proposal-numeric-separator',
+    '@babel/plugin-proposal-throw-expressions',
+    '@babel/plugin-syntax-import-meta',
+  ],
+  presets: [
+    ['@babel/preset-env'],
+    ['@babel/preset-react'],
+  ],
+  test: (fileName) => {
     const nodeModIdx = fileName.lastIndexOf('node_modules');
-    const folioIdx = fileName.lastIndexOf('@folio');
-    if (fileName.endsWith('.js') && (nodeModIdx === -1 || folioIdx > nodeModIdx)) {
-      return false;
+    if (fileName.endsWith('.js') && (nodeModIdx === -1 || fileName.lastIndexOf('@folio') > nodeModIdx)) {
+      return true;
     }
-    return true;
+    return false;
   }
 });