+++ epics/mutationEpics.js
@@ -9,13 +9,19 @@ const actionNames = [
 export default function mutationEpics(resource) {
   return actionNames.map(actionName => (action$) => action$
     .ofType(`@@stripes-connect/${actionName}`)
-    .filter(action => action.meta.resource === resource.name)
+    .filter(action => action.meta.resource === resource.name && !action.meta.silent)
     .map(action => {
       let { meta: { path } } = action;
       path = path && path.replace(/[\/].*$/g, '');  // eslint-disable-line no-useless-escape
 
       const name = resource.name;
-      const meta = Object.assign({}, action.meta, { path, name });
+      const meta = {
+        ...action.meta,
+        originatingActionType: action.type,
+        path,
+        name,
+      };
+
       return { ...action, meta, type: 'REFRESH' };
     }));
 }